requirements.txt
# Core
numpy>=1.23.0
pandas>=1.5.0
PyYAML>=6.0

# PyTorch & Vision
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# OCR
pytesseract>=0.3.10
tesserocr>=2.5.0

# FastAPI backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Model & NLP
transformers>=4.40.0
tokenizers>=0.15.0

# Metrics & Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2

# Utilities
loguru>=0.7.0
python-multipart>=0.0.6

tqdm